Debug & fix stats

+++ nft-marketplace-common/src/main/java/com/tenth/nft/marketplace/common/dao/expression/AbsNftListingUpdate.java
@@ -3,6 +3,8 @@ package com.tenth.nft.marketplace.common.dao.expression;
 import com.tpulse.gs.convention.dao.SimpleUpdate;
 import com.tpulse.gs.convention.dao.annotation.SimpleWriteParam;
 
+import java.math.BigDecimal;
+
 /**
  * @author gs-orm-generator
  * @createdAt 2022/07/05 12:01
@@ -16,7 +18,7 @@ public class AbsNftListingUpdate extends SimpleUpdate {
     private Integer quantity;
 
     @SimpleWriteParam
-    private Float price;
+    private BigDecimal price;
 
     @SimpleWriteParam
     private String currency;
@@ -45,7 +47,7 @@ public class AbsNftListingUpdate extends SimpleUpdate {
         return quantity;
     }
 
-    public Float getPrice(){
+    public BigDecimal getPrice(){
         return price;
     }
 
@@ -83,7 +85,7 @@ public class AbsNftListingUpdate extends SimpleUpdate {
             return this;
         }
 
-        public Builder setPrice(Float price){
+        public Builder setPrice(BigDecimal price){
             update.price = price;
             return this;
         }
